Delete task functionality is added to Task List Component. Buttons added in index.html in Task List Component for other actions with list items. Icons added in Calendar describing that current day has not-resolved tasks. Task List Component is not rerendering now without necessity.

+++ src/fillips/fillipDeleteTask.js
@@ -4,11 +4,28 @@ import { store } from './../index.js';
 
 let fillipDeleteTask = e => {
 
+    var target = e.target;
+
+    const state = store.getState();
+    const currentLocalStorageKey = state.currentLocalStorageKey;
+
+    if (!target.classList.contains('todo__tasks__button_delete')) return; 
+
+    let todaysTasks = JSON.parse(localStorage.getItem(currentLocalStorageKey));
+
+    let newTasks = todaysTasks.filter(function(object) {
+        return target.parentElement.firstChild.innerHTML !== object.content;
+      });
+
+    newTasks.length ? localStorage.setItem(state.currentLocalStorageKey, JSON.stringify(newTasks)) : localStorage.removeItem(state.currentLocalStorageKey);
+
+    console.log(target.parentElement.firstChild.innerHTML);
+
     // console.log('----- Previous state', store.getState());
     store.dispatch({
         type: 'DELETE-TASK-IN-THE-LIST',
         payload: {
-            task: e.target.parentNode.innerHTML,
+            task: newTasks,
         },
     })
     // console.log('+++++ New state', store.getState());
